Created MRP
The new MRP protocol is a protocol designed to sit above TCP connections. This protocol consists of a request, body, and footers. The new package created allows for the user to easily create new MRP requests and parse []byte data for MRP requests. More features will be added to aid in the ease of use for MRP.